Split up cmk.gui.background_job
This is a preparation to move the job modes into this package.
